Adminhtml Download Observer, fix di issues + cleanup
When you are deploying for production you must run the setup:di:compile
and it throws some errors with this module.

The related issue to this Observer is:

Errors during compilation:
    Sebwite\ProductDownloads\Model\Adminhtml\Download\Observer
        Incorrect dependency in class
Sebwite\ProductDownloads\Model\Adminhtml\Download\Observer in
.../vendor/sebwite/magento2-product-downloads/Model/Adminhtml/Download/Observer.php
\Magento\Framework\ObjectManagerInterface already exists in context
object

There were some unused and unprotected parameters in this Observer which
are now cleaned up too.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>